Tag build swift-DEVELOPMENT-SNAPSHOT-2024-05-01-a

+++ lib/SILOptimizer/Utils/CanonicalizeInstruction.cpp
@@ -547,10 +547,19 @@ static SILBasicBlock::iterator
 eliminateUnneededForwardingUnarySingleValueInst(SingleValueInstruction *inst,
                                                 CanonicalizeInstruction &pass) {
   auto next = std::next(inst->getIterator());
-
-  for (auto *use : getNonDebugUses(inst))
-    if (!isa<DestroyValueInst>(use->getUser()))
+  if (isa<DropDeinitInst>(inst))
+    return next;
+  if (auto *uedi = dyn_cast<UncheckedEnumDataInst>(inst)) {
+    if (uedi->getOperand()->getType().isValueTypeWithDeinit())
       return next;
+  }
+  for (auto *use : getNonDebugUses(inst)) {
+    if (auto *destroy = dyn_cast<DestroyValueInst>(use->getUser())) {
+      if (destroy->isFullDeinitialization())
+        continue;
+    }
+    return next;
+  }
   deleteAllDebugUses(inst, pass.callbacks);
   SILValue op = inst->getOperand(0);
   inst->replaceAllUsesWith(op);
